cleanup, and gets user generated logins working
